Changed flock to global shm lock
We're all in on a single machine now, but this should be far
more resilient to running out of memory in the lots of requests
situation. Because that's totally happening :)

This also fixes a bug if exactly the same face
was being built concurrently, and I managed to turn on the logging
and see all the helpful warnings about where I'd forgotten
to write local (curse you, Lua... never again).